Implement the dark/light theme toggler
