+++ src/components/Header.js
@@ -7,9 +7,10 @@ import {
   Toolbar,
   Typography,
 } from '@material-ui/core';
+import PropTypes from 'prop-types';
 import Brightness4Icon from '@material-ui/icons/Brightness4';
 // import Brightness7Icon from '@material-ui/icons/Brightness7';
-import React from 'react';
+import React, { useState } from 'react';
 
 const useStyles = makeStyles(() => ({
   root: {
@@ -24,8 +25,15 @@ const useStyles = makeStyles(() => ({
   },
 }));
 
-const Header = () => {
+const Header = ({ themeFunc }) => {
+  const [bool, setBool] = useState(false);
   const classes = useStyles();
+
+  const themeToggler = () => {
+    setBool(!bool);
+    themeFunc(bool);
+  };
+
   return (
     <AppBar position="static" className={classes.navBackground}>
       <Toolbar>
@@ -34,7 +42,7 @@ const Header = () => {
             <Typography color="primary" variant="h6" className={classes.logoSize}>
               Bookstore CMS
             </Typography>
-            <IconButton color="primary">
+            <IconButton color="primary" onClick={() => themeToggler(bool)}>
               <Brightness4Icon />
             </IconButton>
           </Grid>
@@ -42,6 +50,10 @@ const Header = () => {
       </Toolbar>
     </AppBar>
   );
+};
+
+Header.propTypes = {
+  themeFunc: PropTypes.func.isRequired,
 };
 
 export default Header;
